feat: change layout of MainColumn component
