+++ pkg/web/src/pages/browser/[[...paths]]/components/Revision/index.tsx
@@ -4,9 +4,6 @@ import type { ProjectId, RevisionPath, WorkId } from '@violet/lib/types/branded'
 import type { InfoJson } from '@violet/lib/types/files'
 import { DataIcon } from '@violet/web/src/components/atoms/DataIcon'
 import { Spacer } from '@violet/web/src/components/atoms/Spacer'
-import { CardModal } from '@violet/web/src/components/organisms/CardModal'
-import { useApiContext } from '@violet/web/src/contexts/Api'
-import { useBrowserContext } from '@violet/web/src/contexts/Browser'
 import { colors, fontSizes, mainColumnHeight } from '@violet/web/src/utils/constants'
 import { useEffect, useState } from 'react'
 import styled from 'styled-components'
@@ -19,28 +16,6 @@ const Container = styled.div`
   justify-content: center;
   height: 100%;
 `
-
-const Column = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: end;
-`
-
-const AlertMessage = styled.div`
-  font-size: ${fontSizes.large};
-`
-
-const SecondaryButton = styled.button`
-  width: 108px;
-  height: 36px;
-  font-size: ${fontSizes.large};
-  color: ${colors.white};
-  cursor: pointer;
-  background-color: ${colors.gray};
-  border: none;
-  border-radius: 16px;
-`
-
 const DisplayWorksFrame = styled.div`
   display: flex;
   flex-direction: column;
@@ -88,11 +63,10 @@ export const Revision = (props: {
   projectId: ProjectId
   workId: WorkId
   revision: ApiRevision
+  sendFormData: (file: File) => Promise<void>
+  openAlertModal: (isModal: boolean) => void
 }) => {
   const [isFile, setIsFile] = useState(false)
-  const [openAlert, setOpenAlert] = useState(false)
-  const { api, onErr } = useApiContext()
-  const { wholeDict, updateWholeDict } = useBrowserContext()
   const [workPath, setWorkPath] = useState([props.revision.url])
   const { data, error } = useSWR(props.revision.url, fetcher)
   useEffect(() => {
@@ -108,28 +82,9 @@ export const Revision = (props: {
   }
 
   const dropFile = (file: File) => {
-    fileTypes.some((f) => file.type === f.type) ? void sendFormData(file) : setOpenAlert(true)
-  }
-
-  const sendFormData = async (file: File) => {
-    const revisionRes = await api.browser.projects
-      ._projectId(props.projectId)
-      .works._workId(props.workId)
-      .revisions.$post({ body: { uploadFile: file } })
-      .catch(onErr)
-
-    if (!revisionRes) return
-
-    updateWholeDict('revisionsForWorkId', {
-      [props.workId]: [
-        ...(wholeDict.revisionsForWorkId[props.workId] ?? []),
-        { ...revisionRes, workId: props.workId },
-      ],
-    })
-  }
-
-  const closeModal = () => {
-    setOpenAlert(false)
+    fileTypes.some((f) => file.type === f.type)
+      ? void props.sendFormData(file)
+      : props.openAlertModal(true)
   }
 
   if (error) {
@@ -148,14 +103,6 @@ export const Revision = (props: {
       onDragLeave={() => setIsFile(false)}
       onChange={onChange}
     >
-      <CardModal open={openAlert} onClose={closeModal}>
-        <Column>
-          <Spacer axis="y" size={36} />
-          <AlertMessage>Unsupported file format!</AlertMessage>
-          <Spacer axis="y" size={36} />
-          <SecondaryButton onClick={closeModal}>Confirm</SecondaryButton>
-        </Column>
-      </CardModal>
       {isFile && <Dropper type="file" accept={acceptExtensions} />}
       <DisplayWorksFrame>
         {workPath.map((p, i) => (
